# snapshot (ticdc): reduce list tables time consumption